+++ to_the_STAR/Assets/Game/Scripts/View.cs.meta
