+++ to_the_STAR/Assets/Scenes/SampleScene.unity
@@ -122,7 +122,7 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!43 &446551814
+--- !u!43 &379623391
 Mesh:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -437,6 +437,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 24f6d9ce7227fa440bb9d7235fcba9f5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  mapSizeMin: {x: -5, y: -5}
+  mapSizeMax: {x: 5, y: 5}
 --- !u!1001 &526850365
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -505,7 +507,7 @@ PrefabInstance:
     - target: {fileID: -7204946878765611921, guid: 2fcbc2c357aac8647a346aaffa873480, type: 3}
       propertyPath: m_Mesh
       value: 
-      objectReference: {fileID: 446551814}
+      objectReference: {fileID: 379623391}
     - target: {fileID: 1641597042745260861, guid: 2fcbc2c357aac8647a346aaffa873480, type: 3}
       propertyPath: m_LocalPosition.x
       value: -0.03
